# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Financial Anomaly Detection API",
    "description": "Backend foundation for Machine Learning Enhanced Graph-Based Anomaly Detection for Financial Transactions. Provides transaction ingestion, graph construction, and behavioral/temporal feature extraction.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/analytics/{dataset_id}/users": {
      "get": {
        "tags": [
          "Analytics & Features"
        ],
        "summary": "User Features List",
        "description": "Returns the fused structural (graph), behavioral (transaction statistics), and temporal (frequency, intervals) features calculated for users in the specified dataset. Prepared as tabular features ready for downstream anomaly detection models.",
        "operationId": "get_user_analytics_api_v1_analytics__dataset_id__users_get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter for a specific user ID",
              "title": "User Id"
            },
            "description": "Filter for a specific user ID"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Page size limit",
              "default": 50,
              "title": "Limit"
            },
            "description": "Page size limit"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Page offset index",
              "default": 0,
              "title": "Offset"
            },
            "description": "Page offset index"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAnalyticsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/{dataset_id}/users/{user_id}": {
      "get": {
        "tags": [
          "Analytics & Features"
        ],
        "summary": "Single User Feature Record",
        "description": "Retrieve all structural, behavioral, and temporal features for a single entity in a dataset.",
        "operationId": "get_single_user_features_api_v1_analytics__dataset_id__users__user_id__get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserFeatures"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UserAnalyticsResponse": {
        "properties": {
          "total_users": {
            "type": "integer",
            "title": "Total Users",
            "description": "Total number of evaluated users"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "description": "Page size limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset",
            "description": "Page offset"
          },
          "users": {
            "items": {
              "$ref": "#/components/schemas/UserFeatures"
            },
            "type": "array",
            "title": "Users",
            "description": "List of user feature records"
          }
        },
        "type": "object",
        "required": [
          "total_users",
          "limit",
          "offset",
          "users"
        ],
        "title": "UserAnalyticsResponse",
        "description": "Paginated user feature collection response."
      },
...
      "UserFeatures": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "Unique entity/user identifier"
          },
          "in_degree": {
            "type": "integer",
            "title": "In Degree",
            "description": "Number of incoming transaction connections"
          },
          "out_degree": {
            "type": "integer",
            "title": "Out Degree",
            "description": "Number of outgoing transaction connections"
          },
          "total_degree": {
            "type": "integer",
            "title": "Total Degree",
            "description": "Sum of incoming and outgoing connections"
          },
          "weighted_in_degree": {
            "type": "number",
            "title": "Weighted In Degree",
            "description": "Total volume received across all incoming edges"
          },
          "weighted_out_degree": {
            "type": "number",
            "title": "Weighted Out Degree",
            "description": "Total volume sent across all outgoing edges"
          },
          "betweenness_centrality": {
            "type": "number",
            "title": "Betweenness Centrality",
            "description": "Node betweenness centrality score in the network"
          },
          "transaction_count": {
            "type": "integer",
            "title": "Transaction Count",
            "description": "Total number of transactions involving this user"
          },
          "total_sent": {
            "type": "number",
            "title": "Total Sent",
            "description": "Cumulative monetary amount sent"
          },
          "total_received": {
            "type": "number",
            "title": "Total Received",
            "description": "Cumulative monetary amount received"
          },
          "net_flow": {
            "type": "number",
            "title": "Net Flow",
            "description": "Net flow = total_received - total_sent"
          },
          "average_transaction_amount": {
            "type": "number",
            "title": "Average Transaction Amount",
            "description": "Average amount across all user transactions"
          },
          "maximum_transaction_amount": {
            "type": "number",
            "title": "Maximum Transaction Amount",
            "description": "Maximum amount in a single transaction"
          },
          "unique_receivers": {
            "type": "integer",
            "title": "Unique Receivers",
            "description": "Number of distinct counterparties sent to"
          },
          "unique_senders": {
            "type": "integer",
            "title": "Unique Senders",
            "description": "Number of distinct counterparties received from"
          },
          "transactions_per_day": {
            "type": "number",
            "title": "Transactions Per Day",
            "description": "Average frequency of transactions per day"
          },
          "transactions_per_week": {
            "type": "number",
            "title": "Transactions Per Week",
            "description": "Average frequency of transactions per week"
          },
          "average_time_between_transactions": {
            "type": "number",
            "title": "Average Time Between Transactions",
            "description": "Average duration (seconds) between successive transactions"
          },
          "minimum_time_between_transactions": {
            "type": "number",
            "title": "Minimum Time Between Transactions",
            "description": "Minimum duration (seconds) between successive transactions"
          },
          "maximum_time_between_transactions": {
            "type": "number",
            "title": "Maximum Time Between Transactions",
            "description": "Maximum duration (seconds) between successive transactions"
          },
          "feature_vector": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feature Vector",
            "description": "Raw ordered numerical feature vector for ML model ingestion"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "in_degree",
          "out_degree",
          "total_degree",
          "weighted_in_degree",
          "weighted_out_degree",
          "betweenness_centrality",
          "transaction_count",
          "total_sent",
          "total_received",
          "net_flow",
          "average_transaction_amount",
          "maximum_transaction_amount",
          "unique_receivers",
          "unique_senders",
          "transactions_per_day",
          "transactions_per_week",
          "average_time_between_transactions",
          "minimum_time_between_transactions",
          "maximum_time_between_transactions"
        ],
        "title": "UserFeatures",
        "description": "Unified user profile feature set for downstream anomaly detection."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}